\name{glmc-internal}
\title{Internal glmc Objects}
%
% These are internal
%
\alias{glmc.getXlevels}
%
% These need to be documented!
%
\alias{el.CSW}
\alias{el.owen}
\alias{glmc.inner2}
\alias{flo}
% etc for the others

\description{Internal glmc functions.}
\details{Most of these are not to be called by the user.}

\seealso{glmc-package}
\keyword{internal}
